fix Mana Usage chart (#7225)
this bug is caused by the refactor to `HealingValue`. the refactor itself is fine,
but the use of `any` in this chart hid the use of the value and quietly broke it.

the actual fix is the change to the `.add(...)` call, the rest is just replacing `any`
with actual types.